requirements.txt
amqp==5.2.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.4
Brlapi==0.8.3
celery==5.4.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cron-descriptor==1.4.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
prompt_toolkit==3.0.47
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.7
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zipp==1.0.0
